feat(Player): 사용자 시도 횟수 증가 기능 구현

+++ src/test/java/bridge/PlayerTest.java
@@ -48,6 +48,6 @@ class PlayerTest {
         playerController.increaseMovedDistance();
         Player player = playerRepository.get();
 
-        assertThat(player.getMovedDistance().toInt()).isEqualTo(1);
+        assertThat(player.getMovedDistance()).isEqualTo(1);
     }
 }
